refactor: move EventCard into it's own barrel
