+++ src/components/molecules/EventCard/EventCard.stories.tsx
@@ -1,4 +1,4 @@
-import EventCard from "../EventCard";
+import EventCard from "./EventCard";
 
 export default {
   title: "molecules/EventCard",
