+++ src/components/molecules/EventCard/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./EventCard";
